+++ javascripts/notifications.js
@@ -1,32 +1,43 @@
-// FeedbackGhost is a little class that displays unified feedback from Netzke components.
-Ext.define('Netzke.FeedbackGhost', {
-  showFeedback: function(msg, options){
-    options = options || {};
-    options.delay = options.delay || Netzke.Core.FeedbackDelay;
-    if (Ext.isObject(msg)) {
-      this.msg(msg.level.camelize(), msg.msg, options.delay);
-    } else if (Ext.isArray(msg)) {
-      Ext.each(msg, function(m) { this.showFeedback(m); }, this);
-    } else {
-      this.msg(null, msg, options.delay); // no header for now
-    }
-  },
+/**
+ * Class creating simple notifications. Borrowed from http://dev.sencha.com/extjs/5.1.0/examples/shared/examples.js
+ * @class Netzke.Notifier
+ */
+Ext.define('Netzke.Notifier', function(){
+  var msgCt;
 
-  msg: function(title, format, delay){
-      if(!this.msgCt){
-          this.msgCt = Ext.core.DomHelper.insertFirst(document.body, {id:'msg-div'}, true);
-      }
-      var s = Ext.String.format.apply(String, Array.prototype.slice.call(arguments, 1));
-      var m = Ext.core.DomHelper.append(this.msgCt, this.createBox(title, s), true);
-      m.hide();
-      m.slideIn('t').ghost("t", { delay: delay, remove: true});
-  },
-
-  createBox: function(t, s){
-    if (t) {
-      return '<div class="msg"><h3>' + t + '</h3><p>' + s + '</p></div>';
-    } else {
-      return '<div class="msg"><p>' + s + '</p></div>';
-    }
+  function createBox(t, s){
+    return t ?
+      '<div class="msg ' + Ext.baseCSSPrefix + 'border-box"><h3>' + t + '</h3><p>' + s + '</p></div>'
+    :
+      '<div class="msg ' + Ext.baseCSSPrefix + 'border-box"><p>' + s + '</p></div>';
   }
+
+
+  return {
+    /**
+     * Shows notification on the screen.
+     * @method msg
+     * @param {String} msg Notification body HTML
+     * @param {Object} options May contain the following keys:
+     *
+     *   * **title** - title of notification
+     *   * **delay** (ms) - time notification should stay on the screen
+     */
+    msg: function(msg, options){
+      if (options == undefined) options = {};
+
+      if (Ext.isArray(msg)) {
+        msg = msg.join("<br>")
+      }
+
+      if (msgCt) {
+        document.body.appendChild(msgCt.dom);
+      } else {
+        msgCt = Ext.DomHelper.append(document.body, {id:'msg-div'}, true);
+      }
+      var m = Ext.DomHelper.append(msgCt, createBox(options.title, msg), true);
+      m.hide();
+      m.slideIn('t').ghost("t", { delay: options.delay || Netzke.Core.NotificationDelay, remove: true});
+    }
+  };
 });