+++ lib/netzke/core/services.rb
@@ -140,7 +140,7 @@ module Netzke::Core
     # Called when the method_missing tries to processes a non-existing component. Override when needed.
     # Note: this should actually never happen unless you mess up with Netzke component loading mechanisms.
     def component_missing(missing_component, *params)
-      client.netzke_feedback "Unknown component '#{missing_component}' in '#{name}'"
+      client.netzke_notify "Unknown component '#{missing_component}' in '#{name}'"
     end
 
     private