+++ spec/rails_app/app/components/notifier/client/notifier.js
@@ -1,0 +1,13 @@
+{
+  handleNotify: function(){
+    this.netzkeNotify('Local' + ' feedback', {title: 'Local' + ' notification'});
+  },
+
+  handleServerNotify: function(){
+    this.server.notify();
+  },
+
+  handleMultipleNotify: function(){
+    this.netzkeNotify(['Line' + ' one', 'Line' + ' two']);
+  }
+}